requirements.txt
click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
psycopg2-binary